# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Movie Review Application API
  version: 1.0.0
  description: |
    RESTful API for the Movie Review Application.
    All state-changing endpoints require JWT authentication via httpOnly cookie.
    Admin-only endpoints additionally require the `isAdmin` claim in the JWT payload.

    **Authentication:** JWT access token is delivered as an httpOnly cookie named `access_token`.
    The frontend never accesses the token value directly. On 401 responses, the client
    should POST /auth/refresh to obtain a new access token.

    **CSRF Protection:** All state-changing requests (POST, PUT, PATCH, DELETE) must include
    the `X-CSRF-Token` header with the value read from the non-httpOnly `csrf_token` cookie.

    **Rate Limits:** Limit headers are returned on all responses:
    - `X-RateLimit-Limit`: request limit for the current window
    - `X-RateLimit-Remaining`: remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    **Pagination:** List endpoints return a standard envelope:
    ```json
    { "data": [...], "meta": { "page": 1, "limit": 20, "total": 150, "totalPages": 8 } }
    ```

  contact:
    email: rawaldinesh13@dsu.ac.kr
  license:
    name: MIT
servers:
  - url: /api
    description: Application backend (proxied via Nginx at /api)

# ─────────────────────────────────────────────────────────────────────────────
# SECURITY SCHEMES
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored in httpOnly cookie. Set automatically by the browser
        after successful OAuth login. Expires in 15 minutes. Use POST /auth/refresh
        to renew silently.

  # ─────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ─────────────────────────────────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        totalPages:
          type: integer
          minimum: 0
          example: 8

    # ── Error ────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: "Validation failed"
        error:
          type: string
          example: "Bad Request"
        requestId:
          type: string
          format: uuid
          description: Correlation ID for log tracing

    # ── User ────────────────────────────────────────────────────────────────
    UserPublic:
      type: object
      description: Public user profile (no PII beyond display name and avatar)
      required: [id, displayName, avatarUrl, createdAt]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          example: "Dinesh Rawal"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: "https://lh3.googleusercontent.com/a/..."
        createdAt:
          type: string
          format: date-time
          example: "2026-05-23T10:00:00Z"
    UserMe:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          required: [email, isAdmin]
          properties:
            email:
              type: string
              format: email
              description: Not displayed in public UI
              example: "rawaldinesh13@dsu.ac.kr"
            isAdmin:
              type: boolean
              example: false

    # ── Genre ────────────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, user, movieId, body, isHidden, flagCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserPublic'
        movieId:
          type: string
          format: uuid
        body:
          type: string
          maxLength: 500
          example: "A brutal, poetic film that stays with you long after the credits roll."
        isHidden:
          type: boolean
          description: If true, review is hidden (admin action); only visible to admins
          example: false
        flagCount:
          type: integer
          minimum: 0
          example: 0
        isEdited:
          type: boolean
          description: True if updatedAt > createdAt
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-05-23T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-23T12:00:00Z"
    ReviewCreateRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          description: Review text. Must not be blank or whitespace-only.
          example: "A brutal, poetic film that stays with you long after the credits roll."
    ReviewUpdateRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          example: "Updated review text here."
...
    Rating:
      type: object
      required: [id, userId, movieId, value, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        value:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RatingUpsertRequest:
      type: object
      required: [value]
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 5
          description: Star rating (1-5). Replaces any existing rating by this user for this movie.
          example: 5

    # ── Admin ────────────────────────────────────────────────────────────────
...
  parameters:
    movieId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal movie UUID
    reviewId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal review UUID
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal user UUID
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page

  # ─────────────────────────────────────────────────────────────────────────
  # REUSABLE RESPONSES
  # ─────────────────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: "Unauthorized"
            error: "Unauthorized"
    Forbidden:
      description: Authenticated but insufficient role (not admin)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            message: "Forbidden resource"
            error: "Forbidden"
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: "Resource not found"
            error: "Not Found"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 422
            message: ["body must not be empty", "body must be shorter than or equal to 500 characters"]
            error: "Unprocessable Entity"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 429
            message: "ThrottlerException: Too Many Requests"
    Conflict:
      description: Resource already exists (e.g., duplicate review)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            message: "You have already submitted a review for this movie"
            error: "Conflict"

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /movies/{id}/reviews:
    parameters:
      - $ref: '#/components/parameters/movieId'

    get:
      tags: [Reviews]
      summary: List reviews for a movie
      description: |
        Returns all publicly visible reviews (is_hidden=false, deleted_at=null) for a movie,
        sorted newest-first. Paginated.
        Admins also see hidden reviews (indicated by isHidden=true in the response).
      operationId: listReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Reviews]
      summary: Create review for a movie
      description: |
        Submits a new text review for a movie. One review per user per movie.
        Returns 409 if the user has already submitted a review.
        Rate limited: 20 requests/hour/user.
        Requires authentication (FR-050).
      operationId: createReview
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /reviews/{id}:
    parameters:
      - $ref: '#/components/parameters/reviewId'

    put:
      tags: [Reviews]
      summary: Edit own review
      description: |
        Updates the body of the authenticated user's review.
        Only the review author can edit their own review.
        Returns 403 if the requesting user is not the review author.
        `updatedAt` is set to now(); the edit timestamp is displayed in the UI.
        Requires authentication (FR-053).
      operationId: updateReview
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateRequest'
      responses:
        '200':
          description: Review updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Reviews]
      summary: Delete review (own or admin)
      description: |
        Permanently deletes a review.
        Review authors can delete their own reviews (FR-054).
        Admins can delete any review (FR-071).
        Admin deletions are recorded in the audit log.
        Returns 403 if the requester is neither the author nor an admin.
      operationId: deleteReview
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Review deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /reviews/{id}/flag:
    parameters:
      - $ref: '#/components/parameters/reviewId'

    post:
      tags: [Reviews]
      summary: Flag a review as inappropriate
      description: |
        Allows a registered user to flag a review. One flag per user per review.
        Increments `reviews.flag_count`. Flagged reviews appear in admin moderation queue.
        Returns 409 if the user has already flagged this review (FR-074).
      operationId: flagReview
      security:
        - cookieAuth: []
      responses:
        '201':
          description: Review flagged
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review flagged for moderation"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [Reviews]
      summary: Remove own flag from a review
      description: Allows a registered user to remove their own flag from a review.
      operationId: unflagReview
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Flag removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── RATINGS ────────────────────────────────────────────────────────────────
  /movies/{id}/rating:
    parameters:
      - $ref: '#/components/parameters/movieId'

    post:
      tags: [Ratings]
      summary: Submit or update rating for a movie
      description: |
        Upserts a star rating (1-5) for the authenticated user on a specific movie.
        If the user has already rated this movie, the existing rating is replaced.
        Triggers recalculation of `movies.avg_rating` and `movies.rating_count` via DB trigger.
        Requires authentication (FR-040, FR-041).
      operationId: upsertRating
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingUpsertRequest'
      responses:
        '200':
          description: Rating submitted or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  rating:
                    $ref: '#/components/schemas/Rating'
                  movieAvgRating:
                    type: number
                    format: float
                    description: Updated aggregate rating for the movie
                    example: 4.3
                  movieRatingCount:
                    type: integer
                    description: Updated total rating count
                    example: 248
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Ratings]
      summary: Delete own rating for a movie
      description: |
        Removes the authenticated user's rating for a movie.
        Triggers recalculation of the movie's aggregate rating.
        Returns 404 if the user has not rated this movie (FR-043).
      operationId: deleteRating
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Rating deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users/me:
    get:
      tags: [Users]
      summary: Get own profile
      description: Returns the authenticated user's full profile including email and admin status.
      operationId: getMe
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Users]
      summary: Delete own account (GDPR erasure)
      description: |
        Permanently deletes the authenticated user's account and ALL associated data:
        reviews, ratings, refresh tokens, and flags.
        Executed in a single DB transaction.
        On success: clears all auth cookies and returns 204.
        The user must confirm by sending `X-Confirm-Delete: true` header (FR-062).
      operationId: deleteMe
      security:
        - cookieAuth: []
      parameters:
        - name: X-Confirm-Delete
          in: header
          required: true
          schema:
            type: string
            enum: ["true"]
          description: Safety confirmation header
      responses:
        '204':
          description: Account and all data permanently deleted
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
              description: Clears all auth cookies
        '400':
          description: Missing confirmation header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/userId'

    get:
      tags: [Users]
      summary: Get public user profile
      description: |
        Returns a user's public profile: display name, avatar, member since date,
        and their list of submitted reviews with movie links.
        Does not expose email. Does not expose reviews by deleted users.
        (FR-060, FR-061).
      operationId: getUserProfile
      security: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Public user profile with reviews
          content:
            application/json:
              schema:
                type: object
                required: [user, reviews, meta]
                properties:
                  user:
                    $ref: '#/components/schemas/UserPublic'
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...